+++ reptar/calculators/drivers.py
@@ -20,6 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+from abc import ABC, abstractmethod
 import numpy as np
 from ..utils import chunk_iterable
 from ..logger import ReptarLogger
@@ -32,11 +33,46 @@ except ImportError:
     pass
 
 
-class DriverEnergy:
-    r"""Supervisor of energy+gradient workers.
+def add_worker(
+    workers_list, worker, chunker, worker_args, worker_kwargs, n_cpus_per_worker=1
+):
+    """Add a ray worker to a running list.
 
-    Creates and manages ray tasks using specified worker.
+    Parameters
+    ----------
+    workers_list : :obj:`list`
+        List of all running ray workers.
+    worker : ``callable``
+        Ray worker.
+    chunker : ``iterable``
+        Provides a chunk of indices assigned to this worker.
+    worker_args : :obj:`list`
+        Positional arguments for worker with the exception of ``idxs`` (provided by
+        ``chunker``).
+    worker_kwargs : :obj:`dict`
+        Keyword arguments for worker.
+    n_cpus_per_worker : :obj:`int`, default: ``1``
+        Number of CPU cores to assign the worker.
+
+    Returns
+    -------
+    :obj:`list`
+        Updated ``workers_list``.
     """
+    try:
+        chunk = list(next(chunker))
+        workers_list.append(
+            worker.options(num_cpus=n_cpus_per_worker).remote(
+                chunk, *worker_args, **worker_kwargs
+            )
+        )
+    except StopIteration:
+        pass
+    return workers_list
+
+
+class Driver(ABC):
+    r"""Template driver class"""
 
     def __init__(
         self,
@@ -94,20 +130,155 @@ class DriverEnergy:
                 ray.init(address=ray_address)
 
     @staticmethod
-    def _idx_todo(E, start_slice=None, end_slice=None):
-        r"""Indices of missing energies (calculations to do).
+    @abstractmethod
+    def idx_todo(args, start_slice=None, end_slice=None):
+        r"""Determine which calculations need to be done.
 
-        Determines this by finding all ``NaN`` elements.
+        Parameters
+        ----------
+        args : :obj:`tuple`
+            ``run()`` arguments determined by the specific driver.
+        start_slice : :obj:`int`, default: ``None``
+            Slice arrays in ``args`` starting at this index.
+        end_slice : :obj:`int`, default: ``None``
+            Slice arrays` in ``args`` stopping at this index.
 
         Returns
         -------
         :obj:`numpy.ndarray`
-            Indices for ``R`` that are missing energy values.
+            Indices of source data where calculations need to be ran.
         """
-        # pylint: disable-next=invalid-name
-        E = E[start_slice:end_slice]
-        idx_todo = np.argwhere(np.isnan(E))[:, 0]
+
+    @abstractmethod
+    def check_run_args(self, args):
+        r"""Perform any checks on ``run()`` arguments."""
+
+    @abstractmethod
+    def prep_worker_args(self, args):
+        r"""Prepare worker arguments."""
+
+    @abstractmethod
+    def setup_results(self, args):
+        r"""Setup attribute containing the results of ``run()``."""
+
+    @abstractmethod
+    def process_worker_returns(self, returns):
+        r"""Process worker returns and store in results."""
+
+    @abstractmethod
+    def prepare_run_returns(self):
+        r"""Prepare and process final ``run()`` returns"""
+
+    def _run(self, *args, saver=None):
+        self.check_run_args(args)
+        self.saver = saver
+
+        worker = self.worker
+        idxs_todo = self.idx_todo(args, self.start_slice, self.end_slice)
+        chunker = chunk_iterable(idxs_todo, self.chunk_size)
+        worker_args = self.prep_worker_args(args)
+        self.setup_results(args)
+
+        if not self.use_ray:
+            for idx in idxs_todo:
+                # pylint: disable-next=invalid-name
+                worker_returns = worker([idx], *worker_args, **self.worker_kwargs)
+                self.process_worker_returns(worker_returns)
+        else:
+            worker = ray.remote(worker)
+
+            # Initialize ray workers
+            workers = []
+
+            for _ in range(self.n_workers):
+                workers = add_worker(
+                    workers,
+                    worker,
+                    chunker,
+                    worker_args,
+                    self.worker_kwargs,
+                    self.n_cpus_per_worker,
+                )
+
+            # Start calculations
+            while len(workers) != 0:
+                done_id, workers = ray.wait(workers)
+
+                worker_returns = ray.get(done_id)[0]  # pylint: disable=invalid-name
+                self.process_worker_returns(worker_returns)
+
+                workers = add_worker(
+                    workers,
+                    worker,
+                    chunker,
+                    worker_args,
+                    self.worker_kwargs,
+                    self.n_cpus_per_worker,
+                )
+
+        return self.prepare_run_returns()
+
+
+class DriverEnergy(Driver):
+    r"""Supervisor of energy workers."""
+
+    def __init__(
+        self,
+        worker,
+        worker_kwargs,
+        use_ray=False,
+        n_workers=1,
+        n_cpus_per_worker=1,
+        chunk_size=50,
+        start_slice=None,
+        end_slice=None,
+        ray_address="auto",
+    ):
+        Driver.__init__(
+            self,
+            worker,
+            worker_kwargs,
+            use_ray,
+            n_workers,
+            n_cpus_per_worker,
+            chunk_size,
+            start_slice,
+            end_slice,
+            ray_address,
+        )
+
+    @staticmethod
+    def idx_todo(args, start_slice=None, end_slice=None):
+        # args = (Z, R, E)
+        idx_todo = np.argwhere(np.isnan(args[2][start_slice:end_slice]))[:, 0]
         return idx_todo
+
+    def check_run_args(self, args):
+        # args = (Z, R, E)
+        assert args[1].ndim == 3
+        assert args[0].shape[0] == args[1].shape[1]
+        assert args[1].shape[0] == args[2].shape[0]
+        assert args[1].shape[2] == 3
+
+    def prep_worker_args(self, args):
+        # args = (Z, R, E)
+        if self.use_ray:
+            return ray.put(args[0]), ray.put(args[1])
+        return args[0], args[1]
+
+    def setup_results(self, args):
+        # args = (Z, R, E)
+        self.results = {"E": args[2]}
+
+    def process_worker_returns(self, returns):
+        # returns = (idx_done, E_done)
+        self.results["E"][returns[0]] = returns[1]
+
+        if self.saver is not None:
+            self.saver.save(self.results["E"])
+
+    def prepare_run_returns(self):
+        return self.results["E"]
 
     def run(self, Z, R, E, saver=None):
         r"""Run the calculations.
@@ -135,71 +306,11 @@ class DriverEnergy:
         :obj:`numpy.ndarray`
             The total electronic energy array, ``E``, after all computations.
         """
-        # Check arrays (obsessively)
-        assert R.ndim == 3
-        assert Z.shape[0] == R.shape[1]
-        assert R.shape[0] == E.shape[0]
-        assert R.shape[2] == 3
-
-        worker = self.worker
-        idxs_todo = self._idx_todo(E, self.start_slice, self.end_slice)
-        chunker = chunk_iterable(idxs_todo, self.chunk_size)
-
-        if not self.use_ray:
-            for idx in idxs_todo:
-                # pylint: disable-next=invalid-name
-                _, E_done = worker([idx], Z, R, **self.worker_kwargs)
-                print(E_done)
-                E[idx] = E_done
-
-                if saver is not None:
-                    saver.save(E)
-        else:
-            worker = ray.remote(worker)
-            Z = ray.put(Z)
-            R = ray.put(R)
-
-            # Initialize ray workers
-            workers = []
-
-            def add_worker(workers, chunker):
-                try:
-                    chunk = list(next(chunker))
-                    workers.append(
-                        worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, Z, R, **self.worker_kwargs
-                        )
-                    )
-                except StopIteration:
-                    pass
-
-            for _ in range(self.n_workers):
-                add_worker(workers, chunker)
-
-            # Start calculations
-            while len(workers) != 0:
-                done_id, workers = ray.wait(workers)
-
-                idx_done, E_done = ray.get(done_id)[0]  # pylint: disable=invalid-name
-                E[idx_done] = E_done
-
-                if saver is not None:
-                    saver.save(E)
-
-                add_worker(workers, chunker)
-
-            # Cleanup ray object store
-            del Z
-            del R
-
-        return E
+        return Driver._run(self, Z, R, E, saver=saver)
 
 
-class DriverEnGrad:
-    r"""Supervisor of energy+gradient workers.
-
-    Creates and manages ray tasks using specified worker.
-    """
+class DriverEnGrad(Driver):
+    r"""Supervisor of energy+gradient workers."""
 
     def __init__(
         self,
@@ -213,65 +324,56 @@ class DriverEnGrad:
         end_slice=None,
         ray_address="auto",
     ):
-        r"""
-        Parameters
-        ----------
-        worker : ``callable``
-            The desired worker function to compute energy and gradients. It should
-            be the same as any previous calculations. The ``ray.remote`` decorator
-            will be applied automatically.
-        worker_kwargs : :obj:`tuple`, ndim: ``1``
-            The other keyword arguments for the worker function after
-            ``R_idxs``, ``Z``, and ``R``.
-        use_ray : :obj:`bool`, default: ``False``
-            Use ray to parallelize calculations. If ``False``, calculations are
-            done serially. ``False`` can be useful when running locally or only
-            a few calculations are needed. ``True`` is useful for tons of
-            calculations.
-        n_workers : :obj:`int`, default: ``1``
-            Number of parallel workers to use if ``use_ray`` is ``True``.
-        n_cpus_per_worker : :obj:`int`, default: ``1``
-            Number of CPU cores to provide each worker.
-        chunk_size : :obj:`int`, default: ``50``
-            Number of calculations per task to do. This should be enough to make
-            the ray task overhead significantly less than calculations.
-        start_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to start at this index.
-        end_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to end at this index.
-        ray_address : :obj:`str`, default: ``'auto'``
-            Ray cluster address to connect to.
-        """
-        # Storing arrays and other information
-        self.worker = worker
-        self.worker_kwargs = worker_kwargs
-        self.start_slice = start_slice
-        self.end_slice = end_slice
-
-        self.n_workers = n_workers
-        self.n_cpus_per_worker = n_cpus_per_worker
-        self.chunk_size = chunk_size
-
-        self.use_ray = use_ray
-        if use_ray:
-            if not ray.is_initialized():
-                ray.init(address=ray_address)
+        Driver.__init__(
+            self,
+            worker,
+            worker_kwargs,
+            use_ray,
+            n_workers,
+            n_cpus_per_worker,
+            chunk_size,
+            start_slice,
+            end_slice,
+            ray_address,
+        )
 
     @staticmethod
-    def _idx_todo(E, start_slice=None, end_slice=None):
-        r"""Indices of missing energies (calculations to do).
-
-        Determines this by finding all ``NaN`` elements.
-
-        Returns
-        -------
-        :obj:`numpy.ndarray`
-            Indices for ``R`` that are missing energy values.
-        """
-        # pylint: disable-next=invalid-name
-        E = E[start_slice:end_slice]
-        idx_todo = np.argwhere(np.isnan(E))[:, 0]
+    def idx_todo(args, start_slice=None, end_slice=None):
+        # args = (Z, R, E, G)
+        idx_todo = np.argwhere(np.isnan(args[2][start_slice:end_slice]))[:, 0]
         return idx_todo
+
+    def check_run_args(self, args):
+        # args = (Z, R, E, G)
+        assert args[1].ndim == 3
+        assert args[0].shape[0] == args[1].shape[1]
+        assert args[0].shape[0] == args[3].shape[1]
+        assert args[1].shape[0] == args[3].shape[0]
+        assert args[1].shape[0] == args[3].shape[0]
+        assert args[1].shape[2] == 3
+        assert args[3].shape[2] == 3
+
+    def prep_worker_args(self, args):
+        # args = (Z, R, E, G)
+
+        if self.use_ray:
+            return ray.put(args[0]), ray.put(args[1])
+        return args[0], args[1]
+
+    def setup_results(self, args):
+        # args = (Z, R, E, G)
+        self.results = {"E": args[2], "G": args[3]}
+
+    def process_worker_returns(self, returns):
+        # returns = (idx_done, E_done, G_done)
+        self.results["E"][returns[0]] = returns[1]
+        self.results["G"][returns[0]] = returns[2]
+
+        if self.saver is not None:
+            self.saver.save(self.results["E"], self.results["G"])
+
+    def prepare_run_returns(self):
+        return self.results["E"], self.results["G"]
 
     def run(self, Z, R, E, G, saver=None):
         r"""Run the calculations.
@@ -303,78 +405,13 @@ class DriverEnGrad:
         :obj:`numpy.ndarray`
             The total electronic energy array, ``E``, after all computations.
         :obj:`numpy.ndarray`
-            The atomic gradients array, ``E``, after all computations.
+            The atomic gradients array, ``G``, after all computations.
         """
-        # Check arrays (obsessively)
-        assert R.ndim == 3
-        assert Z.shape[0] == R.shape[1]
-        assert Z.shape[0] == G.shape[1]
-        assert R.shape[0] == E.shape[0]
-        assert R.shape[0] == G.shape[0]
-        assert R.shape[2] == 3
-        assert G.shape[2] == 3
-
-        worker = self.worker
-        idxs_todo = self._idx_todo(E, self.start_slice, self.end_slice)
-        chunker = chunk_iterable(idxs_todo, self.chunk_size)
-
-        if not self.use_ray:
-            for idx in idxs_todo:
-                # pylint: disable-next=invalid-name
-                _, E_done, G_done = worker([idx], Z, R, **self.worker_kwargs)
-                E[idx] = E_done
-                G[idx] = G_done
-
-                if saver is not None:
-                    saver.save((E, G))
-        else:
-            worker = ray.remote(worker)
-            Z = ray.put(Z)
-            R = ray.put(R)
-
-            # Initialize ray workers
-            workers = []
-
-            def add_worker(workers, chunker):
-                try:
-                    chunk = list(next(chunker))
-                    workers.append(
-                        worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, Z, R, **self.worker_kwargs
-                        )
-                    )
-                except StopIteration:
-                    pass
-
-            for _ in range(self.n_workers):
-                add_worker(workers, chunker)
-
-            # Start calculations
-            while len(workers) != 0:
-                done_id, workers = ray.wait(workers)
-
-                # pylint: disable-next=invalid-name
-                idx_done, E_done, G_done = ray.get(done_id)[0]
-                E[idx_done] = E_done
-                G[idx_done] = G_done
-
-                if saver is not None:
-                    saver.save(E, G)
-
-                add_worker(workers, chunker)
-
-            # Cleanup ray object store
-            del Z
-            del R
-
-        return E, G
+        return Driver._run(self, Z, R, E, G, saver=saver)
 
 
-class DriverOpt:
-    r"""Supervisor of optimization workers.
-
-    Creates and manages ray tasks using specified worker.
-    """
+class DriverOpt(Driver):
+    r"""Supervisor of optimization workers."""
 
     def __init__(
         self,
@@ -383,66 +420,58 @@ class DriverOpt:
         use_ray=False,
         n_workers=1,
         n_cpus_per_worker=1,
-        chunk_size=1,
+        chunk_size=50,
         start_slice=None,
         end_slice=None,
         ray_address="auto",
     ):
-        r"""
-        Parameters
-        ----------
-        worker : ``callable``
-            The desired worker function to compute energy and gradients. It should
-            be the same as any previous calculations. The ``ray.remote`` decorator
-            will be applied automatically.
-        worker_kwargs : :obj:`tuple`, ndim: ``1``
-            The other keyword arguments for the worker function after
-            ``R_idxs``, ``Z``, and ``R``.
-        use_ray : :obj:`bool`, default: ``False``
-            Use ray to parallelize calculations. If ``False``, calculations are
-            done serially. ``False`` can be useful when running locally or only
-            a few calculations are needed. ``True`` is useful for tons of
-            calculations.
-        n_workers : :obj:`int`, default: ``1``
-            Number of parallel workers to use if ``use_ray`` is ``True``.
-        n_cpus_per_worker : :obj:`int`, default: ``1``
-            Number of CPU cores to provide each worker.
-        chunk_size : :obj:`int`, default: ``1``
-            Number of calculations per task to do. This should be enough to make
-            the ray task overhead significantly less than calculations.
-        start_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to start at this index.
-        end_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to end at this index.
-        ray_address : :obj:`str`, default: ``'auto'``
-            Ray cluster address to connect to. Passes ``conv_opt, R_opt, E, G`` into
-            :meth:`~reptar.Saver.save`.
-        """
-        # Storing arrays and other information
-        self.worker = worker
-        self.worker_kwargs = worker_kwargs
-        self.start_slice = start_slice
-        self.end_slice = end_slice
-
-        self.n_workers = n_workers
-        self.n_cpus_per_worker = n_cpus_per_worker
-        self.chunk_size = chunk_size
-
-        self.use_ray = use_ray
-        if use_ray:
-            if not ray.is_initialized():
-                ray.init(address=ray_address)
+        Driver.__init__(
+            self,
+            worker,
+            worker_kwargs,
+            use_ray,
+            n_workers,
+            n_cpus_per_worker,
+            chunk_size,
+            start_slice,
+            end_slice,
+            ray_address,
+        )
 
     @staticmethod
-    def _idx_todo(conv_opt, start_slice=None, end_slice=None):
-        r"""Indices of NaN geometries (calculations to do).
+    def idx_todo(args, start_slice=None, end_slice=None):
+        # args = (Z, R, conv_opt, R_opt, E_opt)
+        return np.where(~args[2][start_slice:end_slice])[0]
 
-        Returns
-        -------
-        :obj:`numpy.ndarray`
-            Indices for ``R`` that are missing energy values.
-        """
-        return np.where(~conv_opt[start_slice:end_slice])[0]
+    def check_run_args(self, args):
+        # args = (Z, R, conv_opt, R_opt, E_opt)
+        assert args[1].ndim == 3
+        assert args[0].shape[0] == args[1].shape[1]
+        assert args[1].shape[2] == 3
+
+    def prep_worker_args(self, args):
+        # args = (Z, R, conv_opt, R_opt, E_opt)
+        if self.use_ray:
+            return ray.put(args[0]), ray.put(args[1])
+        return args[0], args[1]
+
+    def setup_results(self, args):
+        # args = (Z, R, conv_opt, R_opt, E_opt)
+        self.results = {"conv_opt": args[2], "R_opt": args[3], "E_opt": args[4]}
+
+    def process_worker_returns(self, returns):
+        # returns = (idx_done, conv_opt_done, R_opt_done, E_opt_one)
+        self.results["conv_opt"][returns[0]] = returns[1]
+        self.results["R_opt"][returns[0]] = returns[2]
+        self.results["E_opt"][returns[0]] = returns[3]
+
+        if self.saver is not None:
+            self.saver.save(
+                self.results["conv_opt"], self.results["R_opt"], self.results["E_opt"]
+            )
+
+    def prepare_run_returns(self):
+        return self.results["conv_opt"], self.results["R_opt"], self.results["E_opt"]
 
     # pylint: disable-next=invalid-name
     def run(self, Z, R, conv_opt, R_opt, E_opt, saver=None):
@@ -473,66 +502,101 @@ class DriverOpt:
         Returns
         -------
         :obj:`numpy.ndarray`
-            The total electronic energy array, ``E``, after all computations.
+            If the geometry optimization converged, ``conv_opt``.
         :obj:`numpy.ndarray`
-            The atomic gradients array, ``E``, after all computations.
+            The last step of geometry optimization coordinates, ``R_opt``.
+        :obj:`numpy.ndarray`
+            The last electronic energy of the geometry optimization, ``E_opt``.
         """
-        # Check arrays (obsessively)
-        assert R.ndim == 3
-        assert Z.shape[0] == R.shape[1]
-        assert R.shape[2] == 3
+        return Driver._run(self, Z, R, conv_opt, R_opt, E_opt, saver=saver)
 
-        worker = self.worker
-        idxs_todo = self._idx_todo(conv_opt, self.start_slice, self.end_slice)
-        chunker = chunk_iterable(idxs_todo, self.chunk_size)
 
-        if not self.use_ray:
-            for idx in idxs_todo:
-                # pylint: disable-next=invalid-name
-                _, conv_opt_done, R_opt_done, E_opt_done = worker(
-                    [idx], Z, R, **self.worker_kwargs
-                )
-                conv_opt[idx] = conv_opt_done
-                R_opt[idx] = R_opt_done[0]
-                E_opt[idx] = E_opt_done[0]
+class DriverCube(Driver):
+    r"""Supervisor of cube workers."""
 
-                if saver is not None:
-                    saver.save(conv_opt, R_opt, E_opt)
-        else:
-            worker = ray.remote(worker)
-            Z = ray.put(Z)
-            R = ray.put(R)
+    def __init__(
+        self,
+        worker,
+        worker_kwargs,
+        use_ray=False,
+        n_workers=1,
+        n_cpus_per_worker=1,
+        chunk_size=50,
+        start_slice=None,
+        end_slice=None,
+        ray_address="auto",
+    ):
+        Driver.__init__(
+            self,
+            worker,
+            worker_kwargs,
+            use_ray,
+            n_workers,
+            n_cpus_per_worker,
+            chunk_size,
+            start_slice,
+            end_slice,
+            ray_address,
+        )
 
-            # Initialize ray workers
-            workers = []
+    @staticmethod
+    def idx_todo(args, start_slice=None, end_slice=None):
+        # args = (Z, R, cube_R, cube_V)
+        # Checks the first element per structure in cube_V for NaN
+        return np.argwhere(np.isnan(args[3][start_slice:end_slice][:, 0]))[:, 0]
 
-            def add_worker(workers, chunker):
-                try:
-                    chunk = list(next(chunker))
-                    workers.append(
-                        worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, Z, R, **self.worker_kwargs
-                        )
-                    )
-                except StopIteration:
-                    pass
+    def check_run_args(self, args):
+        # args = (Z, R, cube_R, cube_V)
+        assert args[1].ndim == 3
+        assert args[0].shape[0] == args[1].shape[1]
+        assert args[1].shape[2] == 3
 
-            for _ in range(self.n_workers):
-                add_worker(workers, chunker)
+    def prep_worker_args(self, args):
+        # args = (Z, R, cube_R, cube_V)
+        if self.use_ray:
+            return ray.put(args[0]), ray.put(args[1]), args[3].shape[-1]
+        return args[0], args[1], args[3].shape[-1]
 
-            # Start calculations
-            while len(workers) != 0:
-                done_id, workers = ray.wait(workers)
+    def setup_results(self, args):
+        # args = (Z, R, cube_R, cube_V)
+        self.results = {"cube_R": args[2], "cube_V": args[3]}
 
-                # pylint: disable-next=invalid-name
-                idxs_done, conv_opt_done, R_opt_done, E_done = ray.get(done_id)[0]
-                conv_opt[idxs_done] = conv_opt_done
-                R_opt[idxs_done] = R_opt_done
-                E_opt[idxs_done] = E_done
+    def process_worker_returns(self, returns):
+        # returns = (idx_done, cube_R, cube_V)
+        self.results["cube_R"][returns[0]] = returns[1]
+        self.results["cube_V"][returns[0]] = returns[2]
 
-                if saver is not None:
-                    saver.save(conv_opt, R_opt, E_opt)
+        if self.saver is not None:
+            self.saver.save(self.results["cube_R"], self.results["cube_V"])
 
-                add_worker(workers, chunker)
+    def prepare_run_returns(self):
+        return self.results["cube_R"], self.results["cube_V"]
 
-        return conv_opt, R_opt, E_opt
+    # pylint: disable-next=invalid-name
+    def run(self, Z, R, cube_R, cube_V, saver=None):
+        r"""Run the calculations.
+
+        Parameters
+        ----------
+        Z : :obj:`numpy.ndarray`, ndim: ``1``
+            Atomic numbers of the atoms with respect to ``R``. Should have shape
+            ``(i,)`` where ``i`` is the number of atoms in the system.
+        R : :obj:`numpy.ndarray`, ndim: ``3``
+            Cartesian coordinates of all structures in group. Should have shape
+            ``(j, i, 3)`` where ``j`` is the number of structures. Units are in
+            Angstroms.
+        cube_R : :obj:`numpy.ndarray`, ndim: ``3``
+            Cartesian coordinates of points where a property is probed.
+        cube_V : :obj:`numpy.ndarray`, ndim: ``2``
+            Property values at the same Cartesian coordinates as ``cube_R``.
+        saver : :obj:`reptar.Saver`, optional
+            Save data after every worker finishes.
+
+        Returns
+        -------
+        :obj:`numpy.ndarray`
+            Cartesian coordinates of points where a property is probed.
+        :obj:`numpy.ndarray`
+            Property values at the same Cartesian coordinates.
+        """
+        return Driver._run(self, Z, R, cube_R, cube_V, saver=saver)