+++ reptar/parsers/__init__.py
@@ -5,5 +5,13 @@ from .orca_parser import ParserORCA
 from .xtb_parser import ParserXTB
 from .ase_parser import ParserASE
 from .crest_parser import ParserCREST
+from .gaussian_cube import parse_cube
 
-__all__ = ["Parser", "ParserORCA", "ParserXTB", "ParserASE", "ParserCREST"]
+__all__ = [
+    "Parser",
+    "ParserORCA",
+    "ParserXTB",
+    "ParserASE",
+    "ParserCREST",
+    "parse_cube",
+]